Move event loop into ProcessManager::run()
Encapsulate the daemon lifecycle (gRPC server, channel wiring,
select! event loop, shutdown) inside ProcessManager, reducing
main() to a thin entry point.
